Move Terminal class to tools
