+++ package.json
@@ -26,7 +26,8 @@
     "swagger-jsdoc": "5.0.1",
     "swagger-ui-express": "4.1.4",
     "uuid": "^3.4.0",
-    "ws": "7.3.1"
+    "ws": "7.3.1",
+    "yargs": "16.2.0"
   },
   "devDependencies": {
     "body-parser": "1.19.0",
@@ -43,8 +44,7 @@
     "nyc": "^15.1.0",
     "open": "7.1.0",
     "temp": "^0.9.4",
-    "xlsx-to-json-lc": "0.5.0",
-    "yargs": "16.2.0"
+    "xlsx-to-json-lc": "0.5.0"
   },
   "engines": {
     "node": ">=12.13.0"
